design: surface wallet's own mints in restore suggestions (#39)
* design: surface wallet's own mints in restore suggestions

SuggestedMintsSection now accepts walletMints param and renders a
"Your mints" section above the curated list. Wallet mints are filtered
against existingURLs (already staged/restored) and deduplicated against
the curated list so nothing appears twice.

Both RestoreWalletView (SettingsView) and the onboarding restore flow
(OnboardingView) now pass walletManager.mints and render
SuggestedMintsSection unconditionally — always visible regardless of
whether the staged list is empty.

* store existing mints
