+++ CashuWallet/Models/Foundation/Data+Hashing.swift
@@ -1,0 +1,10 @@
+import Foundation
+import CryptoKit
+
+extension Data {
+    /// SHA256 hash of the data
+    func sha256() -> Data {
+        let hash = SHA256.hash(data: self)
+        return Data(hash)
+    }
+}
