+++ CashuWallet/Models/Foundation/WalletNotifications.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension Notification.Name {
+    static let cashuTokenReceived = Notification.Name("cashuTokenReceived")
+    static let cashuTokenClaimed = Notification.Name("cashuTokenClaimed")
+    static let cashuTransactionsUpdated = Notification.Name("cashuTransactionsUpdated")
+}
+
+/// Mint quote information
