+++ CashuWallet/Models/Tokens/TokenTransferModels.swift
@@ -1,0 +1,43 @@
+import Foundation
+
+struct SendTokenResult {
+    let token: String
+    let fee: UInt64
+}
+
+/// Pending token entry - stored when user sends ecash
+struct PendingToken: Codable, Identifiable {
+    var id: String { tokenId }
+    let tokenId: String
+    let token: String
+    let amount: UInt64
+    let fee: UInt64
+    let date: Date
+    let mintUrl: String
+    let memo: String?
+}
+
+/// Pending receive token entry - stored when user chooses "Receive Later"
+struct PendingReceiveToken: Codable, Identifiable {
+    var id: String { tokenId }
+    let tokenId: String
+    let token: String
+    let amount: UInt64
+    let date: Date
+    let mintUrl: String
+}
+
+/// Claimed token entry - stored when a sent token is claimed by recipient
+struct ClaimedToken: Codable, Identifiable {
+    var id: String { tokenId }
+    let tokenId: String
+    let token: String
+    let amount: UInt64
+    let fee: UInt64
+    let date: Date
+    let mintUrl: String
+    let memo: String?
+    let claimedDate: Date
+}
+
+/// Result of restoring proofs from a single mint via NUT-09
